+++ lib/generators/rpush_generator.rb
@@ -25,6 +25,7 @@ class RpushGenerator < Rails::Generators::Base
     end
 
     add_rpush_migration('rpush_2_0_0_updates')
+    add_rpush_migration('rpush_2_1_0_updates')
   end
 
    def copy_config