+++ lib/rpush/client/active_model/apns/binary_notification_validator.rb
@@ -6,7 +6,7 @@ module Rpush
           MAX_BYTES = 2048
 
           def validate(record)
-            return unless record.payload_size > MAX_BYTES
+            return unless record.payload.bytesize > MAX_BYTES
             record.errors[:base] << "APN notification cannot be larger than #{MAX_BYTES} bytes. Try condensing your alert and device attributes."
           end
         end